Actu

+++ 42school/libft/ft_strlcat.c
@@ -1,6 +1,6 @@
 #include "libft.h"
 
-int	ft_strlcat(char *dest, char *src, unsigned int size)
+size_t	ft_strlcat(char *dest, char *src, unsigned int size)
 {
 	unsigned int	i;
 
@@ -14,7 +14,7 @@ int	main()
 {
 	char	*s1 = "Je suis dans la joie";
 	char	*s2 = "Je suis ici";
-	printf("%u\n", ft_strlcat(s1, s2, 20));
+	printf("%zu\n", ft_strlcat(s1, s2, 20));
 	printf("%lu\n", strlcat(s1, s2, 20));
 	printf("%lu\n", strlcat(s1, s2, 0));
 	printf("%d\n", ft_strlen(s1));
